+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + [x] Input material: set autocomplete off
 + [x] Tabs: fix clear targets before recollect, issue #1303
 + [x] ListView: fix attributes observing
++ [x] Notify: fix using custom distance option
 
 ### 4.2.34
 + [x] Sidebar: fix error for shifting content issue #1294